Using Integers

+++ hs/src/AST/Domain.hs
@@ -18,4 +18,5 @@ instance Translate (Domainn Expr) (Domain () Expression) where
     toConjure x = mapM toConjure x
 
 dintRange :: Int -> Int -> Domainn Expr
-dintRange a b = DomainInt [RangeBounded (ELit . EI $ a) (ELit . EI $ b)]
+dintRange a b = DomainInt [RangeBounded (ELit . EI $ fromIntegral a)
+                                        (ELit . EI $ fromIntegral b)]
